add config options to abstract away db and runtime args (#24)
Depends on cleanup_api_calls

Adds a configuration file that helps take care of selecting the DB, as well as taking in args so that the Dockerfile doesn't need to be edited every time you want to run something different.

NOTE: There was already a config.py file in the repo, but I couldn't find where it was used anywhere. I just deleted it entirely, but we can rename it to something else if someone is planning to use it.

I originally had the DB hang off the config object itself, but when that happens the initial import creates all of the DBs and causes errors. Still thinking if there's a nice way to encapsulate the DB generator so that main.py doesn't need to know about it at all.

Fixes #16, #20