Adds Batch Calling (#38)
Depends on #35
Fixes #17

Test Plan: All local tests pass, manually verified that a run of main.py produces seemingly correct data into the DB.

This diff adds batch calling. It's a fairly substantial rewrite of the API, but the component changes are:

Pushing params into the request_data() function so that they are more stateless and can be queued in a batch request.
Pushing sql and config params into the initialize function so it is easily accessible everywhere and doesn't need to be passed as a param so much.
Accumulating all of the batch requests and sending them off all at once to a batch object.
Grouping those batches and adding params for batch size (by query) and page size, with defaults based on our own optimization.
Handling those batches as they come in through the callback.
Add a fake batch object for testing purposes.
Cleaning up logging so that it is not as noisy when it is not in debug mode.
A few notes on implementation:

The callback is a nested method per the patterns seen in Google documentation.
The batch.execute() hangs until all of the callbacks have completed, so that gives the response time to add more calls to the queue in cases where there is another page. This is why the while loop works.