ignore partial data in usage report (#47)
Fixes #41

Instead of deleting the last day of data and rerunning it, this now detects partial data issues and doesn't load them at all. This is because there can actually be multiple days of partial data.

Also added a parameter filter to the usage report, which dramatically speeds up the request.

Note that it only does this if the partial data warning is for Google classroom — there are times where the partial data is only for other applications, and that is ok.

For testing / migration, make sure to manually delete any days with partial data first.