add a flag for pulling all data (#48)
Fixes #42

Adds a --all flag for easily calling all of the endpoints. There's a lot more complexity to detecting if none of the flags have been used, so it is worth talking over the approach there if we want to go down that path(do we do it just for args or for the .env file, and is it more clear to have an opt-out rather than opt-in if we do that).