fix pandas type error (#58)
This fixes an error where an empty description causes the pandas reindex function to create a float64 column for description, which then errors when a string is added to it. Instead, it just converts all columns to the object type (before later converting relevant points to the date type). The object dtype is capable of handling ints, floats, and strings.

This is something of a temporary fix, as the long-term solution if data types matter is to enumerate exactly which data type each column should be.

Fixes #40