Adds retry logic for quota limits (#57)
Now that we have optimized the speed of requests, we are running up against user quota limits (600 requests / minute) on the Google Classroom API. This PR adds retry logic so that it snoozes for 20 seconds and then retries the requests. A few notes:

It uses 20 seconds rather than a full 60 because the quota is a rolling 60 second window, so all that needs to happen is for some requests to fall out the back of that window.
It rewrites the batching logic so that there is a stack and it is always just taking the top X requests off the stack. This has the benefit of also not waiting for all pages of one batch to finish before starting on the next batch, and instead can page through multiple batches together.