add course ID to aliases and fix tests (#94)
Fixes #92

Adds a course ID to Aliases, as well as any other future endpoint that might need it.

This also replaces a hacky implementation of execution by course ID in the testing object so that it actually filters the response object by the course ID in the response. It is not used in the case of aliases because the response doesn't contain course IDs, but it was called to attention when I added the API code and discovered the course IDs weren't matching up between request and solution.