Allow test results to actually trigger github workflow pass/fail (#107)
Fixes #106
Fixes an issue where the Github Action was not properly failing when tests failed. This propagates the test failure to the action itself.

Note that the result has to be stored after the first docker-compose call, because otherwise $? will be overwritten by the results of the second call (which should always be successful).